+++ pkg/api/handlers.go
@@ -190,6 +190,7 @@ func NewQuerierHandler(
 	api := v1.NewAPI(
 		engine,
 		errorTranslateQueryable{queryable}, // Translate errors to errors expected by API.
+		nil,                                // No remote write support.
 		func(context.Context) v1.TargetRetriever { return &querier.DummyTargetRetriever{} },
 		func(context.Context) v1.AlertmanagerRetriever { return &querier.DummyAlertmanagerRetriever{} },
 		func() config.Config { return config.Config{} },